+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix: Use `globalThis` instead of `global` or `window` in case the SDK is used in webworkers.
 - Add: Add batch execution support for transports in core.
 - [BREAKING]: Transport.send now accepts a list of items to be sent instead of a single item.
+- Update: Update Vite version in Demo app
 
 ## 1.0.3 – 1.0.5